ROX-15788: Add SecuredCluster subchart to rhacs-terraform
This will add a SecuredCluster CR to the data plane cluster in order to
monitor the cluster via ACS.  Init bundle is saved in AWS parameter
store/secrets manager under the `secured-cluster` prefix.  The
`SECURED_CLUSTER` prefix is used in the `terraform_cluster.sh` script.

The `SecureCluster` CRD is copied over to the helm chart in order to
allow the CR to be created on the first run of the terraform script on a
new cluster, similar to how the observability subchart works.